+++ test/test-utils/data/settingsFiles.js
@@ -6,6 +6,7 @@ const settingsFile5 = require('./settingsFile-5');
 const settingsFile6 = require('./settingsFile-6');
 const settingsFile7 = require('./settingsFile-7');
 const settingsFile8 = require('./settingsFile-8');
+const settingsFile9 = require('./settingsFile-9');
 
 module.exports = {
   settingsFile1,
@@ -15,5 +16,6 @@ module.exports = {
   settingsFile5,
   settingsFile6,
   settingsFile7,
-  settingsFile8
+  settingsFile8,
+  settingsFile9
 };