+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2020-04-29
+### Changes
+- Update track API to handle duplicate and unique conversions and corresponding changes in `launch` API
+- Update track API to track a goal globally across campaigns with the same `goalIdentififer` and corresponding changes in `launch` API
+```js
+// it will track goal having `goalIdentifier` of campaign having `campaignKey` for the user having `userId` as id. 
+vwoClientInstance.track(campaignKey, goalIdentifier, userId, options);
+
+// it will track goal having `goalIdentifier` of campaigns having `campaignKey1` and `campaignKey2` for the user having `userId` as id. 
+vwoClientInstance.track([campaignKey1, campaignKey2], goalIdentifier, userId, options);
+
+// it will track goal having `goalIdentifier` of all the campaigns
+vwoClientInstance.track(null, goalIdentifier, userId, options);
+//Read more about configuration and usage - https://developers.vwo.com/reference#server-side-sdk-track
+
 ## [1.7.4] - 2020-03-05
 ### Changed
 - Added check in segmentation for handling the scenario in which custom-variable key defined in campaign settings is not passed in APIs and matches regex ".*" is used.